+++ src/UserCrudService.ts
@@ -1,0 +1,49 @@
+import { IOrchestratorApi } from './IOrchestratorApi'
+import { BaseCrudService } from '.'
+import { getArray, getData, putData, postData, deleteData } from './utils'
+import path from 'path'
+import { IUserCrudService } from './Interfaces'
+
+export class UserCrudService extends BaseCrudService implements IUserCrudService {
+  constructor(parent_: IOrchestratorApi) {
+    super(parent_)
+  }
+  findAll(queries?: any, asArray: boolean = true): Promise<Array<any>> {
+    return getArray(this.parent.config, this.parent.accessToken, '/odata/Users', queries, asArray)
+  }
+
+  find(id: number): Promise<any> {
+    return getData(this.parent.config, this.parent.accessToken, `/odata/Users(${id})`)
+  }
+
+  _findByUserName(userName: string): Promise<Array<any>> {
+    return this.findAll({ $filter: `UserName eq '${userName}'` })
+  }
+
+  async findByUserName(userName: string): Promise<any> {
+    const users: any[] = await this._findByUserName(userName)
+    return users[0]
+  }
+
+  update(user: any): Promise<any> {
+    return putData(this.parent.config, this.parent.accessToken, `/odata/Users(${user.Id})`, user)
+  }
+
+  create(user: any): Promise<any> {
+    return postData(this.parent.config, this.parent.accessToken, '/odata/Users', user)
+  }
+
+  delete(id: number): Promise<any> {
+    return deleteData(this.parent.config, this.parent.accessToken, `/odata/Users(${id})`)
+  }
+
+  save2Excel(
+    instances: any[],
+    outputFullPath: string,
+    templateFullPath: string = path.join(__dirname, 'templateUsers.xlsx'), // テンプレファイルは、指定されたファイルか、このソースがあるディレクトリ上のtemplateUntitled.xlsxを使う
+    sheetName = 'Sheet1',
+    applyStyles?: (instances: any[], workbook: any, sheetName: string) => void,
+  ): Promise<void> {
+    return super.save2Excel(instances, outputFullPath, templateFullPath, sheetName, applyStyles)
+  }
+}
