+++ src/utils.ts
@@ -2,7 +2,7 @@ import request from 'request'
 import logger, { httpLogger } from './logger'
 import fs from 'fs'
 import path from 'path'
-import OrchestratorApi from './index'
+import { IOrchestratorApi } from './IOrchestratorApi'
 const XlsxPopulate = require('xlsx-populate')
 // import { Parser } from 'json2csv'
 // import url from 'url'
@@ -317,7 +317,7 @@ export const createFilterStr = async (
     level?: 'INFO' | 'TRACE' | 'WARN' | 'ERROR' | 'FATAL'
     machineName?: string
   },
-  api: OrchestratorApi,
+  api: IOrchestratorApi,
 ): Promise<string[]> => {
   const ret: string[] = []
   if (filters.from) {
@@ -358,7 +358,7 @@ export const createAuditFilterStr = async (
     from?: Date
     to?: Date
   },
-  api: OrchestratorApi,
+  api: IOrchestratorApi,
 ): Promise<string[]> => {
   const ret: string[] = []
